+++ server/src/main/java/com/ones/admin/system/ApiResourceService.java
@@ -333,7 +333,7 @@ public class ApiResourceService {
             SystemMenuMapper menuMapper,
             SystemApiManifestSnapshotMapper manifestSnapshotMapper,
             ObjectMapper objectMapper,
-            @Value("${ones.version:v0.0.53}") String applicationVersion
+            @Value("${ones.version:v0.0.54}") String applicationVersion
     ) {
         this.requestMappingHandlerMapping = requestMappingHandlerMapping;
         this.permissionMapper = permissionMapper;